Playfair Cipher.cpp
1. Generate the key square:
I. The key square is a 5×5 grid of alphabets that acts as the key for encrypting the plaintext. 
Each of the 25 alphabets must be unique and one letter of the alphabet (usually J) is 
omitted from the table (as the table can hold only 25 alphabets). If the plaintext contains 
J, then it is replaced by I.
II. The initial alphabets in the key square are the unique alphabets of the key in the order in 
which they appear followed by the remaining letters of the alphabet in order.
2. Algorithm to encrypt the plain text: The plaintext is split into pairs of two letters (digraphs). 
If there is an odd number of letters, a Z is added to the last letter. 
For example:
 PlainText: "instruments"
 After Split: 'in' 'st' 'ru' 'me' 'nt' 'sz'
 Plain Text: “hello”
 After Split: ‘he’ ‘lx’ ‘lo’
 Plain Text: “helloe”
 AfterSplit: ‘he’ ‘lx’ ‘lo’ ‘ex’