Create Rail fence.cpp
A  rail fence cipher is one in which plaintext symbols are rearranged (i.e., transposed or permuted) 
to produce ciphertext. The method of transposition may be either mathematical or typographical 
in nature
Columnar Transposition: The number of columns is the key information.
To encipher : Plaintext is written horizontally in k columns, and is then transcribed vertically 
columnby-column,
To decipher : Suppose that the length of the ciphertext is n and the key is k. Then the letters will
fill n DIV k full rows, and there will be one partial row at the end with n MOD k letters. 
Transcribing row-by-row will then yield the plaintext.
Example: Encrypt NOTHING IN THE WORLD IS MORE DANGEROUS THAN SINCERE IGNORANCE 
AND CONSCIENTIOUS STUPIDITY with a key of k = 9 columns. Solution: We write the plaintext 
horizontally in 9 columns as follows:
The cipher text is therefore: 

NTSESNDTIOHMRIOCIDTEOONROOIHWRUCANUTIOESENSSYNRDTRCCSGLAHEEITIDNAIAEU
NIGNGNNP.