Create DES.c
This is the one of the most hardest cipher tachnics to understand. But very used before AES comes.

The Data Encryption Standard algorithm is a block cipher algorithm that takes in 64-bit blocks 
of plaintext at a time as input and produces 64-bit blocks of cipher text at a time, using a 48-bit 
key for each input. In block cipher algorithms, the text to be encrypted is broken into ‘blocks’ of 
text, and each block is encrypted separately using the key.