Create RSA.cpp
The RSA algorithm is an asymmetric cryptography algorithm; this means that it uses a 
public key and a private key (i.e two different, mathematically linked keys). As their 
names suggest, a public key is shared publicly, while a private key is secret and must not 
be shared with anyone.

Example: 
Choose p = 3 and q = 11
Compute n = p * q = 3 * 11 = 33
Compute φ(n) = (p - 1) * (q - 1) = 2 * 10 = 20
Choose e such that 1 < e < φ(n) and e and φ (n) are coprime. Let e = 7
Compute a value for d such that (d * e) % φ(n) = 1. One solution is d = 3 [(3 * 7) % 20 = 1]
Public key is (e, n) => (7, 33)
Private key is (d, n) => (3, 33)
The encryption of m = 2 is c = 27 % 33 = 29
The decryption of c = 29 is m = 293 % 33 = 2