Create Diffie-Hellman Key Exchange.cpp
The Diffie-Hellman algorithm is being used to establish a shared secret that can be used for secret 
communications while exchanging data over a public network using the elliptic curve to generate 
points and get the secret key using the parameters. 
For the sake of simplicity and practical implementation of the algorithm, we will consider only 4 
variables, one prime q and α (a primitive root of q) and two private values Xa and Xb.
q and α are both publicly available numbers. Users (say Alice and Bob) pick private values Xa and 
Xb and they generate a key and exchange it publicly. The opposite person receives the key and 
that generates a secret key, after which they have the same secret key to encrypt.

Example:
▪ Alice and bob agrees on a prime number 𝒒 = 𝟐𝟑
▪ 𝜶 = 𝟓 as primitive root of 𝑞
▪ Alice selects a private integer 𝑿𝑨 = 𝟔
▪ Alice computes 𝑌𝐴 = 𝛼
𝑋𝐴 mod 𝑞 => 𝒀𝑨 = 𝟓
𝟔 𝒎𝒐𝒅 𝟐𝟑 = 𝟖
▪ Bob selects a private integer 𝑿𝑩 = 𝟏𝟓
▪ Bob computes 𝑌𝐵 = 𝛼
𝑋𝐵 mod 𝑞 => 𝒀𝑩 = 𝟓
𝟏𝟓 𝒎𝒐𝒅 𝟐𝟑 = 𝟏𝟗
▪ Alice sends 𝑌𝐴 to Bob and Bob sends 𝑌𝐵 to Alice
▪ Alice computes key 𝐾 = (𝑌𝐵)
𝑋𝐴 mod 𝑞 => 𝑲 = (𝟏𝟗)
𝟔 𝒎𝒐𝒅 𝟐𝟑
▪ 𝑲 = 𝟐
▪ Bob computes key 𝐾 = (𝑌𝐴)
𝑋𝐵 mod 𝑞=> 𝑲 = (𝟖)
𝟏𝟓 𝒎𝒐𝒅 𝟐𝟑
▪ 𝑲 = 𝟐